LCD backlight control and charging mode
Implements PWM brightness control for the LCD backlight and adds
a mode for when the battery is charging, based off the startup
safety checks. While in the safety loop, if the D switch is set to
'high' position the LCD is dimmed and the battery voltage displayed.
Since the RF is already off in safety mode this gives a relatively
low battery drain for charging without the need for a master on/off
switch.