Fix startup order
The startup sequence activated the radio before reading the persistent store,
using a default packet rate index of 0. This could cause problems if the rx
was using lock_on_first and the saved settings were for anything other than
the default packet rate index.
The code was changed so that the persistent store is read before starting
the radio.