Send sync packets for pkt rate change
When the rx uses lock-on-first it is impossible to change packet
rate without power cycling the rx. To fix this we can send a sync
packet with the new packet rate before changing the rate on the tx.
Since we might lose the packet OTA we set a 'spam' counter to send
multiple sync packets before making the change.