Added user_config.h for local config
This change moves the settings that are most likely to be changed for each
user/handset into user_config.h which is not under version control. An
example is provided as user_config_template.txt which the user should
copy to user_config.h and edit for their own preferences and hardware.
This way, when updating firmware by pulling from git the user config
won't get overwritten.